Rename packages of adapter and dialect
