+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/schema/loader/physical/dialect/impl/SQLServerDatabaseMetaDataDialectHandlerTest.java
@@ -15,11 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.schema.loader.physical.adapter.handler.dialect;
+package org.apache.shardingsphere.infra.metadata.schema.loader.physical.dialect.impl;
 
-import org.apache.shardingsphere.infra.database.type.dialect.PostgreSQLDatabaseType;
-import org.apache.shardingsphere.infra.metadata.schema.loader.physical.adapter.handler.AbstractDatabaseMetaDataDialectHandlerTest;
+import org.apache.shardingsphere.infra.database.type.DatabaseType;
+import org.apache.shardingsphere.infra.database.type.dialect.SQLServerDatabaseType;
+import org.apache.shardingsphere.infra.metadata.schema.loader.physical.dialect.AbstractDatabaseMetaDataDialectHandlerTest;
 import org.apache.shardingsphere.sql.parser.sql.common.constant.QuoteCharacter;
+import org.junit.Before;
 import org.junit.Test;
 
 import java.sql.SQLException;
@@ -28,25 +30,31 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
 
-public final class PostgreSQLDatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
+public final class SQLServerDatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
+    
+    private DatabaseType sqlServerDatabaseType;
+    
+    @Before
+    public void setUp() {
+        sqlServerDatabaseType = new SQLServerDatabaseType();
+    }
     
     @Test
     public void assertGetSchema() throws SQLException {
         when(getConnection().getSchema()).thenReturn(DATABASE_NAME);
-        String postgresqlSchema = getSchema(new PostgreSQLDatabaseType());
-        assertThat(postgresqlSchema, is(DATABASE_NAME));
+        String actualSQLServerSchema = getSchema(sqlServerDatabaseType);
+        assertThat(actualSQLServerSchema, is(DATABASE_NAME));
     }
     
     @Test
     public void assertFormatTableNamePattern() {
-        String postgresqlTableNamePattern = formatTableNamePattern(new PostgreSQLDatabaseType());
-        assertThat(postgresqlTableNamePattern, is(TABLE_NAME_PATTERN));
+        assertThat(formatTableNamePattern(sqlServerDatabaseType), is(TABLE_NAME_PATTERN));
     }
     
     @Test
     public void assertGetQuoteCharacter() {
-        QuoteCharacter postgresqlQuoteCharacter = getQuoteCharacter(new PostgreSQLDatabaseType());
-        assertThat(postgresqlQuoteCharacter.getStartDelimiter(), is("\""));
-        assertThat(postgresqlQuoteCharacter.getEndDelimiter(), is("\""));
+        QuoteCharacter actualSQLServerQuoteCharacter = getQuoteCharacter(sqlServerDatabaseType);
+        assertThat(actualSQLServerQuoteCharacter.getStartDelimiter(), is("["));
+        assertThat(actualSQLServerQuoteCharacter.getEndDelimiter(), is("]"));
     }
 }
